hw2 done
