+++ hw2/package.json
@@ -1,22 +1,14 @@
 {
+  "name": "v",
   "private": true,
+  "version": "0.0.0",
+  "type": "module",
   "scripts": {
-    "start": "webpack serve --mode=development",
-    "build": "webpack --mode=production"
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview"
   },
   "devDependencies": {
-    "@babel/core": "^7.6.2",
-    "@babel/preset-env": "^7.6.2",
-    "@webpack-cli/serve": "^2.0.5",
-    "babel-core": "^6.26.3",
-    "babel-loader": "^8.0.6",
-    "babel-preset-env": "^1.7.0",
-    "css-loader": "^3.2.0",
-    "html-webpack-plugin": "^5.6.0",
-    "mini-css-extract-plugin": "^2.9.1",
-    "style-loader": "^1.3.0",
-    "webpack": "^5.95.0",
-    "webpack-cli": "^5",
-    "webpack-dev-server": "^5.1.0"
+    "vite": "^5.4.8"
   }
 }
