+++ hw2/src/components/chat.css
@@ -1,0 +1,46 @@
+.chat{
+    display: flex;
+    gap: 10px;
+    padding-left: 5px;
+}
+.chat img{
+    height: 57px;
+}
+.chat .body{
+    display: flex;
+    flex-grow: 1;
+    flex-direction: column;
+    border-bottom: 1px solid #ccc;
+    justify-content: center;
+    padding-right: 10px;
+}
+.chat .top,.chat .bottom{
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+}
+.chat .time{
+    color: #797979;
+}
+.chat .text{
+    display: flex;
+    align-items: center;
+    color: #797979;
+}
+.chat .badge.unread,.badge.read{
+    color: #8e24aa;
+}
+.chat .badge.new, .chat .badge.mention{
+    padding: 3px;
+    border-radius: 25%;
+}
+.chat .badge.new{
+    background-color: #8e24aa;
+    color: white;
+}
+.chat .badge.mention{
+    background-color: #fbd767;
+}
+.chat .badge.mention::before{
+    content:'@ '
+}
