+++ hw2/src/components/compose.css
@@ -1,0 +1,10 @@
+.compose{
+    position: fixed;
+    bottom: 10px;
+    right: 10px;
+    border-radius: 50%;
+    background-color: #fbd767;
+    color: #797979;
+    font-size: 25px;
+    padding: 15px;
+}
