+++ hw2/src/index.html
@@ -5,32 +5,25 @@
     <meta name="viewport"
           content="width=device-width, user-scalable=no, initial-scale=1.0, maximum-scale=1.0, minimum-scale=1.0">
     <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <title>Messanger</title>
-    <link rel="stylesheet"
-          href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:opsz,wght,FILL,GRAD@24,400,0,0"/>
+    <title>Messenger</title>
+    <link rel="stylesheet" href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:opsz,wght,FILL,GRAD@20..48,100..700,0..1,-50..200" />
+    <link rel="stylesheet" href="common.css">
+    <link rel="stylesheet" href="screens/screen.css">
+    <link rel="stylesheet" href="/components/topbar.css">
+    <link rel="stylesheet" href="index.css">
+    <link rel="stylesheet" href="components/chat.css">
+    <link rel="stylesheet" href="components/compose.css">
 </head>
 <body>
 <div class="screen">
     <div class="topbar">
-        <span class="material-symbols-outlined">arrow_back</span>
-        <div class="horizontal">
-            <span class="material-symbols-outlined avatar">person</span>
-            <div class="vertical">
-                <div class="name">Дженнифер</div>
-                <div class="last-seen">была 2 часа назад</div>
-            </div>
-        </div>
-        <div><span class="material-symbols-outlined">search</span>
-            <span class="material-symbols-outlined">more_vert</span></div>
+        <span class="material-symbols-outlined">menu</span>
+        <span class="title">Messenger</span>
+        <span class="material-symbols-outlined">search</span>
     </div>
-    <div class="messages-scroll">
-        <div class="messages"></div>
-    </div>
-    <form class="form" action="/">
-        <textarea class="form-input" name="message-text" placeholder="Введите сообщение"></textarea>
-        <span class="material-symbols-outlined">attachment</span>
-        <span class="material-symbols-outlined" id="send">send</span>
-    </form>
-</div>
+    <div id="chats"></div>
+    <span class="material-symbols-outlined compose">edit</span></div>
+<script src="components/chat.js"></script>
+<script src="index.js"></script>
 </body>
 </html>
