+++ hw2/src/screens/chat.css
@@ -1,0 +1,74 @@
+.material-symbols-outlined {
+    cursor: pointer;
+}
+
+.topbar .avatar {
+    background-color: white;
+    color: #333333;
+    border-radius: 50%;
+}
+
+.topbar .vertical {
+    display: flex;
+    flex-direction: column;
+}
+
+.topbar .last-seen {
+    font-size: 0.8em;
+}
+
+.horizontal {
+    display: flex;
+    align-items: center;
+    gap: 10px;
+}
+
+.messages-scroll {
+    overflow-y: scroll;
+    background-color: aliceblue;
+}
+
+.messages{
+    display: flex;
+    flex-direction: column;
+    justify-content: flex-end;
+    min-height: 100%;
+    /*max-height: 500px;*/
+}
+
+.message {
+    border-radius: 10px;
+    margin: 4px;
+    padding: 4px;
+    background-color: white;
+    align-self: flex-start;
+}
+
+.message.my{
+    align-self: flex-end;
+}
+
+.message .top {
+    display: flex;
+    justify-content: space-between;
+}
+
+.form {
+    display: flex;
+    align-items: center;
+    border: 1px solid rgba(25, 25, 25, 0.32);
+}
+
+input[type=submit] {
+    visibility: collapse;
+}
+
+textarea {
+    border: 0;
+    outline: none;
+    width: calc(100% - 2px);
+    font-family: sans-serif;
+  font-size: 1rem;
+  line-height: 1.5rem;
+  padding: .375rem;
+}
