+++ hw2/src/screens/chat.html
@@ -1,0 +1,41 @@
+<!doctype html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport"
+          content="width=device-width, user-scalable=no, initial-scale=1.0, maximum-scale=1.0, minimum-scale=1.0">
+    <meta http-equiv="X-UA-Compatible" content="ie=edge">
+    <title>Messanger</title>
+    <link rel="stylesheet" href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:opsz,wght,FILL,GRAD@20..48,100..700,0..1,-50..200" />
+    <link rel="stylesheet" href="../common.css">
+    <link rel="stylesheet" href="screen.css">
+    <link rel="stylesheet" href="../components/topbar.css">
+    <link rel="stylesheet" href="chat.css">
+</head>
+<body>
+<div class="screen">
+    <div class="topbar">
+        <a class="material-symbols-outlined" href="..">arrow_back</a>
+        <div class="horizontal">
+            <span class="material-symbols-outlined avatar">person</span>
+            <div class="vertical">
+                <div class="name">Дженнифер</div>
+                <div class="last-seen">была 2 часа назад</div>
+            </div>
+        </div>
+        <div><span class="material-symbols-outlined">search</span>
+            <span class="material-symbols-outlined">more_vert</span></div>
+    </div>
+    <div class="messages-scroll">
+        <div class="messages"></div>
+    </div>
+    <form class="form" action="/">
+        <textarea class="form-input" name="message-text" placeholder="Введите сообщение"></textarea>
+        <span class="material-symbols-outlined">attachment</span>
+        <span class="material-symbols-outlined" id="send">send</span>
+    </form>
+</div>
+    <script src="chat.js"></script>
+
+</body>
+</html>
