+++ hw2/src/screens/screen.css
@@ -1,0 +1,26 @@
+html {
+    box-sizing: border-box;
+    height: 99%;
+}
+
+body {
+    margin: 0;
+    background-color: #fff;
+    color: #333;
+    font-size: 15px;
+    font-family: Helvetica, Arial, sans-serif;
+    font-weight: 400;
+    line-height: 20px;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    height: 100%;
+}
+
+.screen {
+    /*max-width: 320px;*/
+    width: 100%;
+    height: 100%;
+    display: flex;
+    flex-direction: column;
+}
